Improved formatting for exported csv file
The time displayed in the csv file is now in seconds. To do this I added
a getTime method to the AnimalTrack class based off the one Danielle
created in the ManualTrackWindowController. I also created a
getPositionsBySecond method in AnimalTrack and have that one being
called in dataExporter rather than the original getPositons method.
The goal is to eventually have the position displayed in cm rather than
pixels but we need to finish up the calibrator first.