Hey Danielle,
I fixed the calibration by drawing on the pane that wrap the video
canvas. I think it's better to do that because that is basically what we
have before, and drawing on the canvas might conflict with your repaint
method. Since the pane wraps the canvas, we know that they will have the
same size.